Moved components and primary color changes implemented
