+++ src/App.jsx
@@ -5,8 +5,8 @@ import Welcome from "./components/pages/Welcome"
 import Home from "./components/pages/home/Home"
 import About from "./components/pages/About"
 import Contact from "./components/pages/Contact"
-import GeneralView from "./components/sections/GeneralView"
-import Calendar from "./components/sections/calendar/Calendar"
+import GeneralView from "./components/pages/home/GeneralView"
+import Calendar from "./components/pages/home/calendar/Calendar"
 import Settings from "./components/pages/home/settings/Settings"
 import Login from "./components/pages/Login"
 import Register from "./components/pages/Register"
