+++ src/components/pages/home/settings/ThemeSelector.jsx
@@ -14,7 +14,7 @@ export default function ThemeSelector() {
     <section className="mb-8 shadow rounded-lg p-6">
       <h1 className="text-5xl font-semibold mb-4">Personalización</h1>
       <h2 className="text-2xl font-semibold mb-2">Tema de color</h2>
-      <div className="flex bg-primary gap-2 rounded dark:bg-slate-800 p-4 w-fit">
+      <div className="flex gap-2 rounded dark:bg-slate-800 p-4 w-fit">
         <div
           onClick={() => handleThemeChange("#0ea5e9")}
           className="w-8 h-8 bg-blue-500 rounded-full cursor-pointer"></div>
