+++ src/index.css
@@ -169,13 +169,13 @@ button:hover {
 }
 
 input[type="text"] {
-  @apply dark:bg-slate-800 text-white disabled:text-gray-400;
+  @apply dark:bg-slate-800 dark:text-white disabled:text-gray-400;
 }
 input[type="date"] {
-  @apply dark:bg-slate-800 text-white disabled:text-gray-400;
+  @apply dark:bg-slate-800 dark:text-white disabled:text-gray-400;
 }
 input[type="email"] {
-  @apply dark:bg-slate-800 text-white disabled:text-gray-400;
+  @apply dark:bg-slate-800 dark:text-white disabled:text-gray-400;
 }
 
 @media (prefers-color-scheme: light) {
